Create README.md
Sure, here is a professional description of yourself:

I am Muhammad Usama Arshad, a recent graduate of the University of Engineering and Technology with a degree in Computer Science. I am a professional programmer with 2 years of experience in web technologies and C++, Python, and JavaScript.

I am a highly skilled and experienced software developer with a strong understanding of object-oriented programming, data structures, and algorithms. I am also proficient in a variety of programming languages, including C++, Python, and JavaScript.

I am a highly motivated and results-oriented individual with a proven track record of success. I am also a team player with excellent communication and interpersonal skills.

I am confident that I have the skills and experience necessary to be a valuable asset to your team. I am eager to learn and grow, and I am confident that I can make a significant contribution to your company.